requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
attrs==21.2.0
EasyProcess==0.3
Faker==9.0.0
iniconfig==1.1.1
makefun==1.11.3
packaging==21.0
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
PyOpenGL==3.0.2
pyparsing==2.4.7
pyscreenshot==0.4.2
pytest==6.2.5
pytest-steps==1.8.0
python-dateutil==2.8.2
PyVirtualDisplay==0.2.1
pywatchman==1.4.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
wrapt==1.12.1
